+++ massiv/src/Data/Massiv/Core.hs
@@ -34,6 +34,7 @@ module Data.Massiv.Core
   , Scheduler
   , SchedulerWS
   , Comp(Seq, Par, Par', ParOn, ParN)
+  , appComp
   , WorkerStates
   , initWorkerStates
   , module Data.Massiv.Core.Index
@@ -56,3 +57,10 @@ import Data.Massiv.Core.Index
 import Data.Massiv.Core.List
 import Data.Massiv.Core.Exception
 
+
+-- | Append computation strategy using `Comp`'s `Monoid` instance.
+--
+-- @since 0.6.0
+appComp :: (Construct r ix e, Load r ix e) => Comp -> Array r ix e -> Array r ix e
+appComp comp arr = setComp (comp <> getComp arr) arr
+{-# INLINEABLE appComp #-}